refactor(portfolio): extract analytics surface into PortfolioAnalytics mixin
Split the jquantstats-backed reporting surface (stats, plots, reports,
sharpe, snapshot and the _build_data/_data caching) out of the 539-line
Portfolio god-class into a new PortfolioAnalytics mixin in _analytics.py.
Portfolio keeps the data model and derived quantities; the public API is
unchanged (Portfolio composes the mixin).

radon MI for portfolio.py rises 20.48 -> 25.58 (out of the borderline A
band); avg cyclomatic complexity stays A (1.49). All 88 tests pass at 100%
coverage; ty + mypy --strict clean.

Closes #783
