+++ src/cvx/simulator/portfolio.py
@@ -17,20 +17,24 @@ This module provides the Portfolio class, which represents a portfolio of assets
 with methods for calculating various metrics (NAV, profit, drawdown, etc.) and
 analyzing performance. The Portfolio class is typically created by the Builder
 class after a simulation is complete.
+
+The jquantstats-backed statistics, plotting, and reporting surface lives in the
+companion :class:`~cvx.simulator._analytics.PortfolioAnalytics` mixin.
 """
 
 from __future__ import annotations
 
 from dataclasses import dataclass, field
 from datetime import datetime
-from typing import Any
 
 import pandas as pd
 from jquantstats.data import Data
 
+from ._analytics import PortfolioAnalytics
+
 
 @dataclass(frozen=True)
-class Portfolio:
+class Portfolio(PortfolioAnalytics):
     """Represents a portfolio of assets with methods for analysis and visualization.
 
     The Portfolio class is a frozen dataclass (immutable) that represents a portfolio
@@ -106,20 +110,6 @@ class Portfolio:
         missing_assets = self.units.columns.difference(self.prices.columns)
         if not missing_assets.empty:
             raise ValueError(f"`units` contains assets not present in `prices`: {missing_assets.tolist()}")  # noqa: TRY003
-
-    def _build_data(self) -> None:
-        """Build and cache the derived quantstats Data from the portfolio NAV.
-
-        This constructs the returns-based :class:`~jquantstats.data.Data` object
-        used by the reporting and statistics helpers and stores it on the frozen
-        instance's ``_data`` field. It is invoked from ``__post_init__`` after
-        the input validation has passed.
-        """
-        frame = self.nav.pct_change().to_frame().reset_index()
-        frame.columns = ["Date", frame.columns[1]]
-        d = Data.from_returns(returns=frame)
-
-        object.__setattr__(self, "_data", d)
 
     @property
     def index(self) -> list[datetime]:
@@ -402,92 +392,6 @@ class Portfolio:
         """
         return self.equity.apply(lambda x: x / self.nav)
 
-    @property
-    def stats(self) -> Any:
-        """Get statistical analysis data for the portfolio.
-
-        This property provides access to various statistical metrics calculated
-        for the portfolio, such as Sharpe ratio, volatility, drawdowns, etc.
-
-        Returns:
-        -------
-        object
-            An object containing various statistical metrics for the portfolio
-
-        Notes:
-        -----
-        The statistics are calculated by the underlying jquantstats library
-        and are based on the portfolio's NAV time series.
-
-        """
-        return self._data.stats
-
-    @property
-    def plots(self) -> Any:
-        """Get visualization tools for the portfolio.
-
-        This property provides access to various plotting functions for visualizing
-        the portfolio's performance, returns, drawdowns, etc.
-
-        Returns:
-        -------
-        object
-            An object containing various plotting methods for the portfolio
-
-        Notes:
-        -----
-        The plotting functions are provided by the underlying jquantstats library
-        and operate on the portfolio's NAV time series.
-
-        """
-        return self._data.plots
-
-    @property
-    def reports(self) -> Any:
-        """Get reporting tools for the portfolio.
-
-        This property provides access to various reporting functions for generating
-        performance reports, risk metrics, and other analytics for the portfolio.
-
-        Returns:
-        -------
-        object
-            An object containing various reporting methods for the portfolio
-
-        Notes:
-        -----
-        The reporting functions are provided by the underlying jquantstats library
-        and operate on the portfolio's NAV time series.
-
-        """
-        return self._data.reports
-
-    def sharpe(self, periods: int | None = None) -> float:
-        """Calculate the Sharpe ratio for the portfolio.
-
-        The Sharpe ratio is a measure of risk-adjusted return, calculated as
-        the portfolio's excess return divided by its volatility.
-
-        Parameters
-        ----------
-        periods : int, optional
-            The number of periods per year for annualization.
-            For daily data, use 252; for weekly data, use 52; for monthly data, use 12.
-            If None, no annualization is performed.
-
-        Returns:
-        -------
-        float
-            The Sharpe ratio of the portfolio
-
-        Notes:
-        -----
-        The Sharpe ratio is calculated using the portfolio's NAV time series.
-        A higher Sharpe ratio indicates better risk-adjusted performance.
-
-        """
-        return float(self.stats.sharpe(periods=periods)["NAV"])
-
     @classmethod
     def from_cashpos_prices(cls, prices: pd.DataFrame, cashposition: pd.DataFrame, aum: float) -> Portfolio:
         """Create a Portfolio instance from cash positions and prices.
@@ -518,22 +422,3 @@ class Portfolio:
         """
         units = cashposition.div(prices, fill_value=0.0)
         return cls(prices=prices, units=units, aum=aum)
-
-    def snapshot(self, title: str = "Portfolio Summary", log_scale: bool = True) -> Any:
-        """Generate and display a snapshot of the portfolio summary.
-
-        This method creates a visual representation of the portfolio summary
-        using the associated plot functionalities. The snapshot can be
-        configured with a title and whether to use a logarithmic scale.
-
-        Args:
-            title: A string specifying the title of the snapshot.
-                   Default is "Portfolio Summary".
-            log_scale: A boolean indicating whether to display the plot
-                       using a logarithmic scale. Default is True.
-
-        Returns:
-            The generated plot object representing the portfolio snapshot.
-
-        """
-        return self.plots.snapshot(title=title, log_scale=log_scale)
